requirements.txt
streamlit>=1.24.0
opencv-python-headless>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
tensorflow-cpu>=2.13.0
imutils>=0.5.4
plotly>=5.13.0
pandas>=1.5.0
scikit-image>=0.19.0
scipy>=1.9.0
protobuf>=3.20.0,<4.0.0 